Remove Elasticsearch6Mapping and make Elasticsearch7Mapping the base implementation
